requirements.txt
streamlit
pandas
numpy
pulp
mplsoccer
matplotlib
Pillow
requests
google-generativeai
